Added media queries to MovieView and ShowView
When you click on a movie or show on a device with a small screen
the page was very crowded and many elements were too big or small.
This commit removed the poster image on smaller screens and made the font sizes grow
and shrink according to the screen size.