Fixed issue where icon colors wouldn't load for Movies
The state for the MovieView component was being set before the
state (watched, interested, favorite) of the movie was being
gotten. So the icons would render and none of them would be colored
even if the user had previously marked it. This commit makes the
state of the movie gotten before anything else, assuring the color
of the appropriate icon is shown.