Fixed issue when marking a show as watched
The end point /mark_watched/ was mishandling marking a show as watched.
More specifically, when marking a show as watch instead of updating
the user's most recently watched show it was updatin the user's
most recently watched movie. This was was fixed by correcting
which variable was set when marking a show as watched.