Fixed issue where watched icon would not load appropriately
In the MovieView and ShowView components the icons were rendering
before the function that got the movie/show's state was finished
running. So, the icon would never appear green even if that item
had been marked as watched. This commit makes the getState()
function the first function that's called, and all subsuquent
functions called after this function is done executing.