Added field to user profile to detect first login
This is the beginning of the work to allow users to set a profile
picture. This new field is a boolean that has a default value of true,
so when a user logins for the first time we can display a window
where the user can upload a profile picture.