Back end changes for uploading profile pictures
These changes add a first_login boolean to the user_profile model
so that the front end can know if it's the first time a user has
logged in and show the profile pic upload window. The image field
has also changed from an image field to a char field that stores
the link to the firebase storage where the image is.