Fixed coloring issue with Rotten Tomatoes rating
The rotten tomatoes ratings were being passed to the Rating
component without being changed to the format which decides
on which color to display the rating as. This commit reformats
the rating if its a rotten tomatoes rating so that the appropriate
color is picked, then is returned to it's original format to display.