Added the ability to explore site without registering
From the Explore window you could click on movies without being
logged in. This patch checks for a RECAPP token, and if not alters
the way MovieView, ShowView and ActorView are rendered if the token
is not defined. These changes include removing the Icons and showing
the navbar without profile and searchbar.